+++ @xen-orchestra/lite/src/jobs/args.ts
@@ -1,14 +1,9 @@
-import type { XenApiHost, XenApiPbd, XenApiSr, XenApiVm } from '@/libs/xen-api/xen-api.types.ts'
+import type { XenApiPbd, XenApiSr, XenApiVm } from '@/libs/xen-api/xen-api.types.ts'
 import { defineJobArg } from '@core/packages/job'
 
 export const vmsArg = defineJobArg({
   identify: (vm: XenApiVm) => vm.$ref,
   toArray: true,
-})
-
-export const hostArg = defineJobArg({
-  identify: (host: XenApiHost) => host.$ref,
-  toArray: false,
 })
 
 export const srsArg = defineJobArg({
