+++ @xen-orchestra/lite/src/modules/host/jobs/host-args.ts
@@ -1,0 +1,7 @@
+import type { XenApiHost } from '@/libs/xen-api/xen-api.types.ts'
+import { defineJobArg } from '@core/packages/job'
+
+export const hostArg = defineJobArg({
+  identify: (host: XenApiHost) => host?.$ref,
+  toArray: false,
+})
